Machine Learning project template (#2)
Adds a new project type, which is suitable for machine learning (ML). The dependencies are based around pytorch (including `numpy`, `scipy`). Additional dependencies for the ML project type include: `pandas` for data manipulation; `pytorch-lightning` for a rapid model development framework; Hugging Face's `transformers` and `spacy` for Natural Language Processing (NLP); the popular `torchvision` and `albumentations` libraries for Computer Vision (CV); and `tqdm` for progress bars.

Additionally, the ML template includes executable script templates for the ML lifecycle: downloading, formatting, and making training data; training and evaluating a model; and offline, batch prediction and a stub for a predictive webserver.

On a different note, this PR adds the `pywise` library as a dependency for all project templates.

Version `0.1.0`